Add Control tree, XML-RPC command, CPU and memory metrics, Crossbuild.
Introduce daqtree tool and nodetree library
* control tree for daqling components based on configuration
* possibility to include/exclude nodes
* inconsistent flag for children in different states

Reorganize daqcontrol and supervisor_wrapper libraries
* remove print statements from library
* exception are forwarded to the method caller
* introduce concurrent.futures in order to return values from command threads
* add code documentation

XML-RPC Command implementation
* Overhaul Command implementation from ZMQ Req-Rep socket to XML-RPC client.
* Moved Command actions to dedicated XML-RPC methods.
* Cleanup of deprecated methods in Command and ConnectionManager.
* Aligned daqcontrol library to XML-RPC transport.
* Closes #19 .
* Add step for xmlrpc-c installation in Ansible setup-host role (updated registry CI image).

CPU and Memory metrics
* add psutils-manager.py

Crossbuild features
* add Ansible roles for installation of Linaro toolchain
* add setup scripts for cross-compilation of different architectures
* build with CPR made optional

+++ src/Modules/ReadoutInterface/ReadoutInterfaceModule.cpp
@@ -61,7 +61,8 @@ void ReadoutInterfaceModule::runner() {
 
   std::random_device rd;
   std::mt19937 gen(rd());
-  std::uniform_int_distribution<> dis(m_min_payload, m_max_payload);
+  std::uniform_int_distribution<> dis(static_cast<int>(m_min_payload),
+                                      static_cast<int>(m_max_payload));
 
   DEBUG("Running...");
   while (m_run) {
